phishing-detector/requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Web Interface
streamlit>=1.24.0
requests>=2.31.0

# NLP & Text Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# URL Processing
tldextract>=3.4.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Visualization & Analysis (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter (for notebooks)
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.23.0

# Development Tools
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Progress Bars
tqdm>=4.65.0
